+++ k8s-readonly.tf
@@ -42,6 +42,37 @@ resource "kubernetes_deployment_v1" "apiary_hms_readonly" {
       spec {
         service_account_name            = kubernetes_service_account_v1.hms_readonly[0].metadata.0.name
         automount_service_account_token = true
+
+        dynamic "toleration" {
+          for_each = var.hms_ro_tolerations
+          content {
+            effect             = lookup(toleration.value, "effect", null)
+            key                = lookup(toleration.value, "key", null)
+            operator           = lookup(toleration.value, "operator", null)
+            value              = lookup(toleration.value, "value", null)
+          }
+        }
+
+        dynamic "affinity" {
+          for_each = var.hms_ro_node_affinity
+          content {
+            node_affinity {
+              required_during_scheduling_ignored_during_execution {
+                dynamic "node_selector_term" {
+                  for_each = lookup(affinity.value, "node_selector_term", [])
+                  content {
+                    match_expressions {
+                      key      = lookup(node_selector_term.value, "key", null)
+                      operator = lookup(node_selector_term.value, "operator", null)
+                      values   = lookup(node_selector_term.value, "values", [])
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+
         dynamic "security_context"  {
           for_each = var.enable_tcp_keepalive ? ["enabled"] : []
           content {
@@ -59,6 +90,7 @@ resource "kubernetes_deployment_v1" "apiary_hms_readonly" {
             }
           }
         }
+
         dynamic "init_container" {
           for_each = var.external_database_host == "" ? ["enabled"] : []